update paths after reorganisation
