+++ navigation/SpotlightNavigator.js
@@ -4,25 +4,25 @@ import React from 'react';
 import usePopToTopOnBlur from '../hooks/usePopToTopOnBlur';
 import EditProductScreen, {
   screenOptions as editProductScreenOptions,
-} from '../screens/addAndEdit/EditProductScreen';
+} from '../screens/EditProductScreen';
 import EditProfileScreen, {
   screenOptions as editProfileScreenOptions,
-} from '../screens/addAndEdit/EditProfileScreen';
+} from '../screens/EditProfileScreen';
 import EditProjectScreen, {
   screenOptions as editProjectScreenOptions,
-} from '../screens/addAndEdit/EditProjectScreen';
+} from '../screens/EditProjectScreen';
 import EditProposalScreen, {
   screenOptions as editProposalScreenOptions,
-} from '../screens/addAndEdit/EditProposalScreen';
-import ProductDetailScreen from '../screens/details/ProductDetail/ProductDetailScreen';
-import ProjectDetailScreen from '../screens/details/ProjectDetailScreen';
-import ProposalDetailScreen from '../screens/details/ProposalDetailScreen';
-import ProductsScreen from '../screens/shop/ProductsScreen';
-import SpotlightScreen from '../screens/shop/SpotlightScreen';
-import UserProductsScreen from '../screens/user/UserProductsScreen';
-import UserProfile from '../screens/user/UserProfile';
-import UserProposalsScreen from '../screens/user/UserProposalsScreen';
-import UserSpotlightScreen from '../screens/user/UserSpotlightScreen';
+} from '../screens/EditProposalScreen';
+import ProductDetailScreen from '../screens/ProductDetailScreen';
+import ProjectDetailScreen from '../screens/ProjectDetailScreen';
+import ProposalDetailScreen from '../screens/ProposalDetailScreen';
+import ProductsScreen from '../screens/ProductsScreen';
+import SpotlightScreen from '../screens/SpotlightScreen';
+import UserProductsScreen from '../screens/UserProductsScreen';
+import UserProfile from '../screens/UserProfile';
+import UserProposalsScreen from '../screens/UserProposalsScreen';
+import UserSpotlightScreen from '../screens/UserSpotlightScreen';
 import {
   detailHeader,
   defaultNavOptions,
