+++ navigation/TabNavigator.js
@@ -7,20 +7,20 @@ import Colors from '../constants/Colors';
 import usePopToTopOnBlur from '../hooks/usePopToTopOnBlur';
 import EditProductScreen, {
   screenOptions as editProductScreenOptions,
-} from '../screens/addAndEdit/EditProductScreen';
+} from '../screens/EditProductScreen';
 import EditProjectScreen, {
   screenOptions as editProjectScreenOptions,
-} from '../screens/addAndEdit/EditProjectScreen';
+} from '../screens/EditProjectScreen';
 import EditProposalScreen, {
   screenOptions as editProposalScreenOptions,
-} from '../screens/addAndEdit/EditProposalScreen';
-import ProductDetailScreen from '../screens/details/ProductDetail/ProductDetailScreen';
-import ProjectDetailScreen from '../screens/details/ProjectDetailScreen';
-import ProposalDetailScreen from '../screens/details/ProposalDetailScreen';
-import ProductsScreen from '../screens/shop/ProductsScreen';
-import ProjectsScreen from '../screens/shop/ProjectsScreen';
-import ProposalsScreen from '../screens/shop/ProposalsScreen';
-import UserSpotlightScreen from '../screens/user/UserSpotlightScreen';
+} from '../screens/EditProposalScreen';
+import ProductDetailScreen from '../screens/ProductDetailScreen';
+import ProjectDetailScreen from '../screens/ProjectDetailScreen';
+import ProposalDetailScreen from '../screens/ProposalDetailScreen';
+import ProductsScreen from '../screens/ProductsScreen';
+import ProjectsScreen from '../screens/ProjectsScreen';
+import ProposalsScreen from '../screens/ProposalsScreen';
+import UserSpotlightScreen from '../screens/UserSpotlightScreen';
 import { topStackHeaderForTabs, detailHeaderForTabs, mainPageOptionsNoUser } from './NavHeaders';
 import { SpotlightNavigator } from './SpotlightNavigator';
